Rename folders in the absence of index files
