+++ src/lib/Scenes/Consignments/v2/Boot.tsx
@@ -4,7 +4,7 @@ import { ProvideScreenDimensions } from "lib/utils/useScreenDimensions"
 import React, { useEffect, useRef } from "react"
 import { View } from "react-native"
 import { Modal } from "./Components/Modal"
-import { artworkSchema, validateArtworkSchema } from "./Screens/MyCollectionAddArtwork/Form/artworkSchema"
+import { artworkSchema, validateArtworkSchema } from "./Screens/AddArtwork/Form/artworkSchema"
 import { ArtworkFormValues } from "./State/artworkModel"
 import { useStoreActions, useStoreState } from "./State/hooks"
 import { store } from "./State/store"
