+++ src/lib/Scenes/Consignments/v2/Screens/AddArtwork/Components/ArtistSearchResult.tsx
