+++ src/lib/Scenes/Consignments/v2/Screens/AddArtwork/Components/SizePicker.tsx
@@ -1,5 +1,5 @@
 import { Select } from "lib/Components/Select"
-import { useArtworkForm } from "lib/Scenes/Consignments/v2/Screens/MyCollectionAddArtwork/Form/useArtworkForm"
+import { useArtworkForm } from "lib/Scenes/Consignments/v2/Screens/AddArtwork/Form/useArtworkForm"
 import React, { useRef } from "react"
 
 export const SizePicker: React.FC = () => {
