+++ src/lib/Scenes/Consignments/v2/Screens/ArtworkDetail/AuctionResults.tsx
