+++ src/lib/Scenes/Consignments/v2/Screens/ArtworkDetail/MyCollectionArtworkDetail.tsx
